Improvements, added ini file, atten control
Added ini file support.
Changed band pass switching to follow tun setting instead of lo setting.
Added SetAtten command to ExtIO interface to allow control of atten, and rf preamp. Basically if the db agrument is 0 the rf pream and atten is disengaed, if it is positive, then the RF Preamp is engaged, if it is negative the attenutator is turned on.